+++ packages/console/src/components/CreateTenantModal/types.ts
@@ -1,0 +1,3 @@
+import { type TenantModel } from '@logto/schemas';
+
+export type CreateTenantData = Pick<TenantModel, 'name' | 'tag'>;
